+++ config-base-next.js
@@ -4,6 +4,7 @@ module.exports = {
     {
       files: ['**/*.ts', '**/*.tsx'],
       rules: {
+        '@typescript-eslint/no-floating-promises': ['error', { ignoreIIFE: true }],
         '@typescript-eslint/promise-function-async': [
           'error',
           {
@@ -19,6 +20,9 @@ module.exports = {
 
         'unicorn/no-null': 'warn',
       },
+    },
+    {
+      files: ['**/test/**/*'],
     },
   ],
 };